collector-docker/requirements.txt
pycrypto
mysql-connector-python